+++ components/bookCatalog/AddBookModal/index.tsx
@@ -1,0 +1,2 @@
+import AddBookModal from './AddBookModal';
+export default AddBookModal;
